+++ src/db.rs
@@ -14,6 +14,10 @@ const DB_MIGRATION_V2: &str = "mig_2";
 const DB_MIGRATION_V3: &str = "mig_3";
 const DB_MIGRATION_V4: &str = "mig_4";
 const DB_MIGRATION_V5: &str = "mig_5";
+const DB_MIGRATION_V6: &str = "mig_6";
+const DB_MIGRATION_V7: &str = "mig_7";
+const DB_MIGRATION_V8: &str = "mig_8";
+const DB_MIGRATION_V9: &str = "mig_9";
 
 #[derive(Debug)]
 struct ILAgentItem {
@@ -149,6 +153,50 @@ impl ILDatabase {
             info!("Database migrated to {}", DB_MIGRATION_V5);
         }
 
+        let mig_6 = self.get_il_value(DB_MIGRATION_V6);
+        if mig_6.is_none() {
+            self.conn
+                .execute("ALTER TABLE event_items ADD COLUMN labels TEXT NULL", [])
+                .expect("Database migration failed (v6)");
+
+            self.set_il_val(DB_MIGRATION_V6, DB_MIGRATION_VAL)
+                .expect("Database migration failed (v6, set)");
+            info!("Database migrated to {}", DB_MIGRATION_V6);
+        }
+
+        let mig_7 = self.get_il_value(DB_MIGRATION_V7);
+        if mig_7.is_none() {
+            self.conn
+                .execute("ALTER TABLE event_items ADD COLUMN severity INTEGER NULL", [])
+                .expect("Database migration failed (v7)");
+
+            self.set_il_val(DB_MIGRATION_V7, DB_MIGRATION_VAL)
+                .expect("Database migration failed (v7, set)");
+            info!("Database migrated to {}", DB_MIGRATION_V7);
+        }
+
+        let mig_8 = self.get_il_value(DB_MIGRATION_V8);
+        if mig_8.is_none() {
+            self.conn
+                .execute("ALTER TABLE event_items ADD COLUMN routing_key TEXT NULL", [])
+                .expect("Database migration failed (v8)");
+
+            self.set_il_val(DB_MIGRATION_V8, DB_MIGRATION_VAL)
+                .expect("Database migration failed (v8, set)");
+            info!("Database migrated to {}", DB_MIGRATION_V8);
+        }
+
+        let mig_9 = self.get_il_value(DB_MIGRATION_V9);
+        if mig_9.is_none() {
+            self.conn
+                .execute("ALTER TABLE event_items ADD COLUMN services TEXT NULL", [])
+                .expect("Database migration failed (v9)");
+
+            self.set_il_val(DB_MIGRATION_V9, DB_MIGRATION_VAL)
+                .expect("Database migration failed (v9, set)");
+            info!("Database migrated to {}", DB_MIGRATION_V9);
+        }
+
         /*
         Run simple db migrations, if needed, like this:
 
@@ -244,12 +292,16 @@ impl ILDatabase {
             custom_details: row.get(9).unwrap_or(None),
             details: row.get(10).unwrap_or(None),
             event_api_path: row.get(11).unwrap_or(None),
+            labels: row.get(12).unwrap_or(None),
+            severity: row.get(13).unwrap_or(None),
+            routing_key: row.get(14).unwrap_or(None),
+            services: row.get(15).unwrap_or(None),
         })
     }
 
     pub fn get_il_event(&self, event_id: &str) -> Result<Option<EventQueueItem>, rusqlite::Error> {
         let mut stmt = self.conn.prepare("SELECT id, integration_key, event_type, alert_key, summary, created_at,
-         priority, images, links, custom_details, details, event_api_path FROM event_items WHERE id = ?1")?;
+         priority, images, links, custom_details, details, event_api_path, labels, severity, routing_key, services FROM event_items WHERE id = ?1")?;
         let query_result =
             stmt.query_map(&[&event_id], |row| ILDatabase::convert_db_row_to_event(row));
 
@@ -281,7 +333,7 @@ impl ILDatabase {
 
     pub fn get_il_events(&self, limit: i32) -> Result<Vec<EventQueueItem>, rusqlite::Error> {
         let mut stmt = self.conn.prepare("SELECT id, integration_key, event_type, alert_key, summary, created_at,
-         priority, images, links, custom_details, details, event_api_path FROM event_items ORDER BY inserted_at ASC LIMIT ?1")?;
+         priority, images, links, custom_details, details, event_api_path, labels, severity, routing_key, services FROM event_items ORDER BY inserted_at ASC LIMIT ?1")?;
         let query_result =
             stmt.query_map(&[&limit], |row| ILDatabase::convert_db_row_to_event(row));
 
@@ -321,12 +373,13 @@ impl ILDatabase {
 
         let insert_result = self.conn.execute(
             "INSERT INTO event_items (integration_key, event_type, alert_key, summary, created_at, id,
-                priority, images, links, custom_details, details, event_api_path)
-                  VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12)",
+                priority, images, links, custom_details, details, event_api_path,
+                labels, severity, routing_key, services)
+                  VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14, ?15, ?16)",
             &[&item.integration_key as &dyn ToSql, &item.event_type, &item.alert_key,
                 &item.summary, created_at, &item_id,
                 &item.priority, &item.images, &item.links, &item.custom_details, &item.details,
-                &item.event_api_path],
+                &item.event_api_path, &item.labels, &item.severity, &item.routing_key, &item.services],
         );
 
         match insert_result {